requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
pyyaml
faker
matplotlib
pandas
streamlit-option-menu